Mon, Feb 27, 2023, 9:31 PM +06:00

+++ test/learn.c
@@ -17,10 +17,10 @@ int main(void)
     int n;
     scanf("%d", &n);
 
-    int sheep[n], ref[n];
+    long long int sheep[n], ref[n];
     for (int i = 0; i < n; i++)
     {
-        scanf("%d", &sheep[i]);
+        scanf("%lld", &sheep[i]);
     }
 
     for (int i = 0; i < n; i++)
@@ -30,8 +30,8 @@ int main(void)
 
     // sheep[2] --- 5 6
 
-    int attacked = 0;   //
-    int non_stolen = 0; //
+    long long int attacked = 0;   //
+    long long int non_stolen = 0; //
 
     for (int j = 0;;)
     {
@@ -68,7 +68,7 @@ int main(void)
         non_stolen += sheep[i];
     }
 
-    printf("%d %d\n", attacked, non_stolen);
+    printf("%lld %lld\n", attacked, non_stolen);
 
     return 0;
 }
